Renaming with the JSX file extension.
