+++ src/entrypoint.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import Resume from './pages/resume/Resume';
+import Resume from './pages/resume/Resume.jsx';
 
 const app = document.getElementById('app');
